Lab 5 Integrating_RDS_with_EKS/cart/requirements.txt
Flask
flask-cors
psycopg2-binary